+++ ratarmount.py
@@ -44,7 +44,7 @@ except ImportError:
     pass
 
 
-__version__ = '0.8.0'
+__version__ = '0.8.1'
 
 
 parallelization = 1